pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Ramil Albakov <ramil.albakov2@gmail.com>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.3"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.5"
python-dotenv = "^1.0.0"
validators = "^0.20.0"
requests = "^2.28.2"
beautifulsoup4 = "^4.12.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
isort = "^5.12.0"
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
